refactor: page container component folder structure
