+++ .storybook/main.js
@@ -3,8 +3,7 @@ const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
 module.exports = {
   stories: [
     '../shared/**/*.stories.@(js|jsx|ts|tsx)',
-    '../index-page/*.stories.tsx',
-    '../case-study-page/*.stories.tsx'
+    '../page-container/**/*.stories.@(js|jsx|ts|tsx)'
   ],
   addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
   webpackFinal: async (config) => {
