+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import Head from 'next/head';
-import IndexPage from 'pedialab-index/IndexPage';
+import IndexPage from 'pedialab-pages/index-page/IndexPage';
 
 function HomePage() {
   return (
